# blog-openapi.yaml
openapi: "3.0.3"
info:
  title: "Blog API"
  description: "Blog API"
  version: "1.0.0"
servers:
  - url: "https://localhost:8080"
...
paths:
...
  /api/v1/user/update:
    put:
      tags:
        - user
      summary: "PUT api/v1/user/update"
      operationId: "updateUserInfo"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUserInfo"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Object"
...
components:
  schemas:
    Object:
      type: "object"
      properties: { }
    Date:
      type: "object"
      properties: { }
    PostUserInfo:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        username:
          type: "string"
        password:
          type: "string"
        email:
          type: "string"
        phoneNumber:
          type: "string"
        description:
          type: "string"
        sex:
          type: "string"
          enum:
            - "type"
            - "DEFAULT"
            - "MALE"
            - "FEMALE"
        bornDate:
          $ref: "#/components/schemas/Date"
        region:
          type: "string"
        website:
          type: "string"
        photo:
          type: "string"